refactor: move board reaction types to own file
