+++ src/types/reaction.ts
@@ -21,22 +21,3 @@ export const REACTION_EMOJI_MAP = new Map<ReactionType, string>([
   ["celebration", "🥳"],
   ["poop", "💩"],
 ]);
-
-/**
- * representation of a board reaction on the server side
- */
-export type BoardReactionType = {
-  id: string;
-  user: string;
-  reactionType: ReactionType;
-};
-
-export const BOARD_REACTION_EMOJI_MAP = new Map<ReactionType, string>([
-  ["tada", "🎉"],
-  ["applause", "👏"],
-  ["heart", "💖"],
-  ["like", "👍"],
-  ["dislike", "👎"],
-]);
-
-export type BoardReactionState = BoardReactionType[];
